move IrcChannelUser to a separate file
