+++ src/event/irc/EventIrcChatListing.cpp
@@ -7,20 +7,6 @@ UUID EventIrcChatListing::getEventUuid() const {
     return this->uuid;
 }
 
-IrcChannelUser::IrcChannelUser(const std::string& nick,
-                               const std::string& mode)
-    : nick{nick}
-    , mode{mode}
-{
-}
-
-std::string IrcChannelUser::getNick() const {
-    return nick;
-}
-
-std::string IrcChannelUser::getMode() const {
-    return mode;
-}
 
 IrcChannelListing::IrcChannelListing(const std::string& channelName,
                                      const std::string& channelTopic,
